+++ agent/metrics/cpu_darwin.go
@@ -1,0 +1,10 @@
+//go:build darwin
+
+package metrics
+
+import "go.opentelemetry.io/otel/metric"
+
+// addCPUMetrics is not supported for macOS.
+func addCPUMetrics(_ metric.Meter) error {
+	return nil
+}
