+++ agent/otel.go
@@ -5,26 +5,22 @@ import (
 	"context"
 	"encoding/base64"
 	"encoding/hex"
-	"fmt"
 	"os"
 	"path"
 	"time"
 
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/agent/internal"
+	"github.com/evergreen-ci/evergreen/agent/metrics"
 	"github.com/evergreen-ci/utility"
 	"github.com/mongodb/grip"
 	"github.com/pkg/errors"
-	"github.com/shirou/gopsutil/v3/cpu"
-	"github.com/shirou/gopsutil/v3/mem"
-	"github.com/shirou/gopsutil/v3/net"
 	"go.opentelemetry.io/contrib/detectors/aws/ec2"
 	"go.opentelemetry.io/contrib/detectors/aws/ecs"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
-	"go.opentelemetry.io/otel/metric"
 	sdk "go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
@@ -41,20 +37,6 @@ const (
 	packageName    = "github.com/evergreen-ci/evergreen/agent"
 	traceSuffix    = "build/OTelTraces"
 	maxLineSize    = 1024 * 1024
-
-	cpuTimeInstrumentPrefix = "system.cpu.time"
-	cpuUtilInstrument       = "system.cpu.utilization"
-
-	memoryUsageInstrumentPrefix = "system.memory.usage"
-	memoryUtilizationInstrument = "system.memory.utilization"
-
-	diskIOInstrumentPrefix         = "system.disk.io"
-	diskOperationsInstrumentPrefix = "system.disk.operations"
-	diskIOTimeInstrumentPrefix     = "system.disk.io_time"
-
-	networkIOInstrumentPrefix = "system.network.io"
-
-	processCountPrefix = "system.process.count"
 )
 
 func (a *Agent) initOtel(ctx context.Context) error {
@@ -135,138 +117,7 @@ func (a *Agent) startMetrics(ctx context.Context, tc *internal.TaskConfig) (func
 
 	return func(ctx context.Context) {
 		grip.Error(errors.Wrap(meterProvider.Shutdown(ctx), "doing meter provider"))
-	}, errors.Wrap(instrumentMeter(ctx, meterProvider.Meter(packageName)), "instrumenting meter")
-}
-
-func instrumentMeter(ctx context.Context, meter metric.Meter) error {
-	catcher := grip.NewBasicCatcher()
-
-	catcher.Wrap(addCPUMetrics(meter), "adding CPU metrics")
-	catcher.Wrap(addMemoryMetrics(meter), "adding memory metrics")
-	catcher.Wrap(addDiskMetrics(ctx, meter), "adding disk metrics")
-	catcher.Wrap(addNetworkMetrics(meter), "adding network metrics")
-	catcher.Wrap(addProcessMetrics(meter), "adding process metrics")
-
-	return catcher.Resolve()
-}
-
-func addCPUMetrics(meter metric.Meter) error {
-	cpuTimeIdle, err := meter.Float64ObservableCounter(fmt.Sprintf("%s.idle", cpuTimeInstrumentPrefix), metric.WithUnit("s"))
-	if err != nil {
-		return errors.Wrap(err, "making cpu time idle counter")
-	}
-	cpuTimeSystem, err := meter.Float64ObservableCounter(fmt.Sprintf("%s.system", cpuTimeInstrumentPrefix), metric.WithUnit("s"))
-	if err != nil {
-		return errors.Wrap(err, "making cpu time system counter")
-	}
-	cpuTimeUser, err := meter.Float64ObservableCounter(fmt.Sprintf("%s.user", cpuTimeInstrumentPrefix), metric.WithUnit("s"))
-	if err != nil {
-		return errors.Wrap(err, "making cpu time user counter")
-	}
-	cpuTimeSteal, err := meter.Float64ObservableCounter(fmt.Sprintf("%s.steal", cpuTimeInstrumentPrefix), metric.WithUnit("s"))
-	if err != nil {
-		return errors.Wrap(err, "making cpu time steal counter")
-	}
-	cpuTimeIOWait, err := meter.Float64ObservableCounter(fmt.Sprintf("%s.iowait", cpuTimeInstrumentPrefix), metric.WithUnit("s"))
-	if err != nil {
-		return errors.Wrap(err, "making cpu time iowait counter")
-	}
-
-	cpuUtil, err := meter.Float64ObservableGauge(cpuUtilInstrument, metric.WithUnit("1"), metric.WithDescription("Busy CPU time since the last measurement, divided by the elapsed time"))
-	if err != nil {
-		return errors.Wrap(err, "making cpu util gauge")
-	}
-
-	_, err = meter.RegisterCallback(func(ctx context.Context, observer metric.Observer) error {
-		times, err := cpu.TimesWithContext(ctx, false)
-		if err != nil {
-			return errors.Wrap(err, "getting CPU times")
-		}
-		if len(times) != 1 {
-			return errors.Wrap(err, "CPU times had an unexpected length")
-		}
-		observer.ObserveFloat64(cpuTimeIdle, times[0].Idle)
-		observer.ObserveFloat64(cpuTimeSystem, times[0].System)
-		observer.ObserveFloat64(cpuTimeUser, times[0].User)
-		observer.ObserveFloat64(cpuTimeSteal, times[0].Steal)
-		observer.ObserveFloat64(cpuTimeIOWait, times[0].Iowait)
-
-		return nil
-	}, cpuTimeIdle, cpuTimeSystem, cpuTimeUser, cpuTimeSteal, cpuTimeIOWait)
-	if err != nil {
-		return errors.Wrap(err, "registering cpu time callback")
-	}
-
-	_, err = meter.RegisterCallback(func(ctx context.Context, observer metric.Observer) error {
-		util, err := cpu.PercentWithContext(ctx, 0, false)
-		if err != nil {
-			return errors.Wrap(err, "getting CPU util")
-		}
-		if len(util) != 1 {
-			return errors.Wrap(err, "CPU util had an unexpected length")
-		}
-		observer.ObserveFloat64(cpuUtil, util[0])
-
-		return nil
-	}, cpuUtil)
-	return errors.Wrap(err, "registering cpu time callback")
-}
-
-func addMemoryMetrics(meter metric.Meter) error {
-	memoryUsageAvailable, err := meter.Int64ObservableUpDownCounter(fmt.Sprintf("%s.available", memoryUsageInstrumentPrefix), metric.WithUnit("By"))
-	if err != nil {
-		return errors.Wrap(err, "making memory usage available counter")
-	}
-
-	memoryUsageUsed, err := meter.Int64ObservableUpDownCounter(fmt.Sprintf("%s.used", memoryUsageInstrumentPrefix), metric.WithUnit("By"))
-	if err != nil {
-		return errors.Wrap(err, "making memory usage used counter")
-	}
-
-	memoryUtil, err := meter.Float64ObservableGauge(memoryUtilizationInstrument, metric.WithUnit("1"))
-	if err != nil {
-		return errors.Wrap(err, "making memory util gauge")
-	}
-
-	_, err = meter.RegisterCallback(func(ctx context.Context, observer metric.Observer) error {
-		memStats, err := mem.VirtualMemoryWithContext(ctx)
-		if err != nil {
-			return errors.Wrap(err, "getting memory stats")
-		}
-		observer.ObserveInt64(memoryUsageAvailable, int64(memStats.Available))
-		observer.ObserveInt64(memoryUsageUsed, int64(memStats.Used))
-		observer.ObserveFloat64(memoryUtil, memStats.UsedPercent)
-
-		return nil
-	}, memoryUsageAvailable, memoryUsageUsed, memoryUtil)
-	return errors.Wrap(err, "registering memory callback")
-}
-
-func addNetworkMetrics(meter metric.Meter) error {
-	networkIOTransmit, err := meter.Int64ObservableCounter(fmt.Sprintf("%s.transmit", networkIOInstrumentPrefix), metric.WithUnit("by"))
-	if err != nil {
-		return errors.Wrap(err, "making network io transmit counter")
-	}
-
-	networkIOReceive, err := meter.Int64ObservableCounter(fmt.Sprintf("%s.receive", networkIOInstrumentPrefix), metric.WithUnit("by"))
-	if err != nil {
-		return errors.Wrap(err, "making network io receive counter")
-	}
-	_, err = meter.RegisterCallback(func(ctx context.Context, observer metric.Observer) error {
-		counters, err := net.IOCountersWithContext(ctx, false)
-		if err != nil {
-			return errors.Wrap(err, "getting network stats")
-		}
-		if len(counters) != 1 {
-			return errors.Wrap(err, "network counters had an unexpected length")
-		}
-
-		observer.ObserveInt64(networkIOTransmit, int64(counters[0].BytesSent))
-		observer.ObserveInt64(networkIOReceive, int64(counters[0].BytesRecv))
-
-		return nil
-	}, networkIOTransmit, networkIOReceive)
-	return errors.Wrap(err, "registering network io callback")
+	}, errors.Wrap(metrics.InstrumentMeter(ctx, meterProvider.Meter(packageName)), "instrumenting meter")
 }
 
 func hostResource(ctx context.Context) (*resource.Resource, error) {
